Proposed change to allow the user to keep TV show artwork locally
(This is primarily for iOS devices since they only show square covers for TV shows but TVDB only provides portrait-styled images.)

You just need to drop a cover.jpg file into your Sickbeard directories and it will pick them up. I made it 'cover.jpg' since the images provided by iTunes are
JPEG only but I guess this could be changed to a setting or set to look for a PNG as well.