iOS audio behavior change
Changed the behavior of the iOS audio option to avoid creating duplicate
stereo audio channels. Script will now ignore the preferred codec if the
source is only stereo and opt for a single AAC 2.0 channel when the iOS
audio option is enabled. This maximizes mobile compatibility while
minimizing file size.