consolidate usenet post options
Consolidates all the usenet post processing scripts into 2 files, one
for SAB and one for NZBGet. The “category” will determine whether the
files will be passed to Sickbeard, Sonarr, or CouchPotato. Additionally
a new `convert`option has been added to the Sickbeard and CouchPotato
sections of autoProcess.ini to allow early conversion, eliminating the
need for custom “withConverter” editions of the post scripts.