manual.py fix - avoid conversion during move_to if destination exists
In manual.py if move_to is specific and the destination file exists the script will fail.  This adds in a check to skip the conversion if the destination file already exists.  In addition to resource efficient (no dead files left behind), this enables to script to be run multiple times against the same media collection.