ffmpeg output parser fix
The ffmpeg parser expects '\r' in the output to yield to the caller.  For small or fast ffmpeg jobs, ffmpeg never outputs the '\r' so the job is detected as fails.

This patch handles the case where a job is finished but never yielded.  Seems to not mess with legitimate errors.