Fixed original media being left with .original extention
When the ffmpeg convert fails the except block sets outputfile to None. This causes the if statement on line 832 to always fire, and then the os.rename fails. This leaves the original media with the .original extension. This fix will instead rename the file back to what it was and then re-throw the exception (which is now unhandled, to gracefully force the script exit)