valid file scanning overhaul
Remove static old validSource and needProcessing methods and unify under new isValidSource method

isValidSource does not use predefined input and output extensions but instead tries to read the file with ffprobe and returns none if its not a valid format or there are no streams. This info can be passed on to the process or generate options command (or both) so that the file doesn't need to be probed multiple times. This reduces restrictions significantly and allows

Similarly a new isValidSubtitle method to check external subs, again not relying on hardcode extensions

isImagedBasedSubtitle method which uses FFMPEG to try a test conversion and see if it can be converted to text. If this fails this is due to it being image based and will be marked and handled accordingly

Update all the post scripts to use these new validation methods

Fix bug where temp_extension wasn't being applied and certain renaming conditions would cause it to be left behind

Adjust directory walking method used in manual.py to prevent looping over created files during the same loop

Fix edge case delete issues when using combinations of settings

Moved audio default activity to debug level of logging except for final default track report

safePrint method for manual.py to reduce number of try/except blocks

improve external subtitle detection with added flexibility

in an effort to move towards being more extension neutral convert-mp4 has been renamed to process-same-extensions and behavior is more dynamic